Update Reasonable EVA Jetpack for KSP 1.11.1
KSP 1.11.1 fixes free refills for non-default propellants, allowing the use of Nitrogen in lieu of generic stock EVA Propellant.

Added commented-out code to disable free refills in hopes that resource transfer in cargo will eventually be supported. Yet to come: actual storage of Nitrogen in pods and tanks.

Bumped up the capacity of the fuel cylinder.